Add component for a client's credentials
